Bootstrap
Added bootstrap to handle element layouts.
Fixed some meta data.
In index.js, calls are made to WAMP server which is now discoverable on
the internet and not just local. Done as a test to see how I would go
about this if I was to use an ec2 instance.