Added home page and more
Replaced the index page so it is now a landing page with links to the
different requirements of the app, although these themselves are not all
complete.

Added a css file so I can start tweaking some elements when I get to
that stage.

driverSignOut1.html is now what index.html use to be.