Driver sign in layout created
Finished much of the layout for the driver sign in screen.
Still to add is the vehicle number for the selected driver.
The driver name drop down list is not filling with the names, although
results are returned correctly from the query.

Created a JS file, driverSignOut.js, to deal with the handling of
queries in the process of signing out.