Local storage variables added
For the driver sign out screen 1 of 3, local variables are now set each
time a value is changed in the drop down lists. These can then be
retrieved in sign out 3 of 3 to fill in the fields.

The same has been done to sign out screen 2 of 3, where the health and
safety standards are saved to local storage as variables for later use.

Driver sign out 3 of 3 now sets all the elements on the page, using the
values from local storage.

In driverSignOut.js, code for retrieving the PDAs that can be signed out
has been added and is now on screen 1 of 3 of the driver sign out
process.

index.html has a call to this function when "driver sign out" menu
option is pressed.