Setting and getting values from local storage
The variables for various driver sign out choices are saved as
variables. These are checked when returning to a page so that, if a
driver returns to sign out 1 from sign out 2, they remain checked. This
is the same from 2 to 3.
This is achieved through the use of scripts.

Changed the method names to set instead of get, as they are setting the
values.

Added a home button to the top left - needs aligning and linking to
index.html