Removing surplass files
As I rolled a test of my app and the queries into what will be the final
version I had extra files. The test was a web page that called PHP files
to deal with the queries. In the final version, these PHP files will be
hosted server side so are no longer required here.