Fixed admin back button and authing users
The back button on the Admin control page now returns home and clears
the admin name from local stroage.

auth.js takes the value entered in a prompt dialog box and checks the
database for a valid password. If it is found, the name of the user
along with paye number is assigned to localstorage for use in the
add/remove van (when implemented).