Successfully post new van data to the database
The data now posts data to the database successfully.
I assign the paye number to a local storage key and use this to submit
to the database instead of the full name (due to database changes).

Added a plugin for the use of toasts. Will show a success message when
the data is submitted successfully.