Loading of JS moved to onLoad function
Instead of loading the relevant scripts from index.html when an option
is selected, moved them loading to an onLoad function at bottom of page.
This is mostly for the driver signing out process. If the user reaches
page 2 or 3 and needed to return, the lists were not populating again as
the call was only made from the index.html onClick event.