Van list changes & back button work
Had to make changes to code for obtaining van information due to some
backend changes.

Attempting to fix the back button, so it asks user if they want to
cancel when on certain pages.

Removed an uncessary file.

Added alternative row colours to the table in the view log to make it
easier to read.