Mostly UI changes
The elements for collection duty walks now toggles between show/hide
should the user enter the wrong name and then change it. Previously if
the user entered a  name and they were not on a collection duty it would
hide this element and not show it again if they had entered the wrong
name and was on a collection duty.

The variables for collection duty tasks are now, Y/N and NA. This is
looks better in the view log section where the admins can see a simple
Y/N or NA against the columns.

Removed the back button from top of all admin pages. The onBackButton
deals with this, no need for it twice.

The "Back" button on adminIndex is changed to show "log out" to better
represent the action that is taken when pressed.

Applied some CSS to the view log table to make it bit easier to read and
fit screen more comfortably.