Form validation
Changed most pages to incorporate some form validation.

Most attributes that are required are now marked as such and the user
cannot progress until they have checked certain boxes. Form only submits
when the user has done so.

Reworked the submit function of the form and button to manage this.

Changed the radio buttons for yes and no options to just a single
checkbox for each requirement, so user cannot select "no" for correct
footwear for example.

Had to adjust the css of the inline-checkbox to work around some
formatting issue.